# Simplify 3-way split function in RTree bulk load